Add -time-resolvers option for more summary
Use `-time-resolvers` to see a duration summary for each resolver, highlighting
which the slow ones are.  Combines nicely with `-progress`.

It was adding this option (upstream) which revealed the bug fixed in the
previous commit, that I hadn't switched which set of strings we iterate over
when I added flag parsing.