Fix inadvertent lock-step in parallelism
Avoiding buffering in channels meant that for each item to be resolved, we
ended up waiting for all resolutions on the resolvers to finish before
proceeding, such that the total duration would be the sum of the durations of
the slowest duration for each item to be resolved.

With this, roughly, the duration for each resolver is independent and the total
duration is the max of the per-resolver durations.

Noticed, again, thanks to the per-resolver timing stats, which were
suspiciously too similar to each other.

Also, align the stats